refactor: add commands
