+++ data.csv
@@ -1,0 +1,46 @@
+name,total_ops,threads,spent,throughput,latency
+RwLock<BTreeMap>,167772,1,107942839,1554267.0690734752,643
+RwLock<BTreeMap>,167772,2,63499691,2642091.5969496607,756
+RwLock<BTreeMap>,167772,3,52856718,3174090.3776886035,945
+RwLock<BTreeMap>,167772,4,40363160,4156562.568441123,962
+RwLock<BTreeMap>,167772,5,36162906,4639339.548652423,1077
+RwLock<BTreeMap>,167772,6,36771992,4562494.193950657,1315
+RwLock<BTreeMap>,167772,7,34221249,4902567.992185206,1427
+RwLock<BTreeMap>,167772,8,33529122,5003769.558892714,1598
+RwLock<BTreeMap>,167772,9,30864118,5435826.807038517,1655
+RwLock<FxHashMap>,167772,1,14467690,11596322.564279439,86
+RwLock<FxHashMap>,167772,2,16119786,10407830.476161407,192
+RwLock<FxHashMap>,167772,3,19699082,8516742.04919803,352
+RwLock<FxHashMap>,167772,4,25003946,6709820.921865693,596
+RwLock<FxHashMap>,167772,5,17085691,9819444.820815265,509
+RwLock<FxHashMap>,167772,6,16949221,9898508.019926107,606
+RwLock<FxHashMap>,167772,7,12175989,13778921.777935246,508
+RwLock<FxHashMap>,167772,8,18693577,8974847.349974807,891
+RwLock<FxHashMap>,167772,9,23606771,7106944.020425327,1266
+FxDashMap,167772,1,20121482,8337954.430990719,119
+FxDashMap,167772,2,8944503,18756995.218180373,106
+FxDashMap,167772,3,6407519,26183613.345508613,114
+FxDashMap,167772,4,5786490,28993742.32047407,137
+FxDashMap,167772,5,4730350,35467143.023243524,140
+FxDashMap,167772,6,4840699,34658630.91260167,173
+FxDashMap,167772,7,3628115,46242194.63826257,151
+FxDashMap,167772,8,3013791,55668093.77292586,143
+FxDashMap,167772,9,3022591,55506021.158668175,162
+FxFlurry,167772,1,19720593,8507452.083210682,117
+FxFlurry,167772,2,9714223,17270758.55680892,115
+FxFlurry,167772,3,8212636,20428520.14870743,146
+FxFlurry,167772,4,6493493,25836941.68916483,154
+FxFlurry,167772,5,5301372,31646901.97179145,157
+FxFlurry,167772,6,5964096,28130331.906126257,213
+FxFlurry,167772,7,3913696,42867918.20315119,163
+FxFlurry,167772,8,4441416,37774439.503077395,211
+FxFlurry,167772,9,3842457,43662687.70216557,206
+FxEvmap,167772,1,52680089,3184732.6605693474,313
+FxEvmap,167772,2,29182340,5749093.458578031,347
+FxEvmap,167772,3,19192035,8741751.460957631,343
+FxEvmap,167772,4,15141729,11080108.486950202,361
+FxEvmap,167772,5,13758522,12194042.354258692,410
+FxEvmap,167772,6,11983437,14000323.947128024,428
+FxEvmap,167772,7,9135449,18364942.982003402,381
+FxEvmap,167772,8,8898617,18853716.25725661,424
+FxEvmap,167772,9,10369544,16179303.54507392,556
