+++ src/plot.rs
@@ -1,0 +1,6 @@
+use structopt::StructOpt;
+
+#[derive(Debug, StructOpt)]
+pub struct Options {}
+
+pub fn plot(options: &Options) {}
